add check when deleting course of not deleting if there are final inscriptions for it (open or closed) (#135)

+++ services/src/main/java/ar/edu/itba/paw/services/CourseServiceImpl.java
@@ -104,6 +104,10 @@ public class CourseServiceImpl implements CourseService {
       return false;
     }
 
+    if(courseDao.getCourseFinalInscriptions(courseId) != null) {
+      return false;
+    }
+
     deleteCourseCorrelatives(courseId);
 
     courseDao.deleteCourse(courseId);
